[silgen] Try to re-abstract first before going through other code paths in visitRecNonInOutBase. NFC.
The main thing here is that we have a large block of code that we will never hit
if we need to re-abstract. Instead, this commit reflows visitRecNonInOutBase to
first check if re-abstraction is needed before doing anything else. If
re-abstration is required, we quickly handle it and return.

This commit also fixes some variable names to match the rest of the function.

rdar://31521023