[ClangImporter] Bump the maximum supported version to Swift 5
"...finally."

This was technically enabled two commits ago, since nothing checks
that you're not /over/ maxVersion(). This is only used for

- deciding the canonical way to import renamed types
- trying to import things in multiple ways

...and so there are very few observable differences, especially
before anyone has added any API notes that differentiate Swift 4
and Swift 5.

At some point we should start encoding name versions in the lookup
tables so that we only have to try all the names /once/, but the test
suite doesn't seem to get measureably slower with this change,
probably because it's pretty quick to decide that most things don't
have multiple names. So we can put that off until later.