Start implementing a more complete Type::join.
This just refactors the existing code into a TypeVisitor with no new
functionality added.

The intent is that for cases where two different TypeKinds have a
join, the lower-numbered type kind will contain the implementation and
the other will dispatch to that case. I'm not far enough along to know
how well this is going to work out yet.