ASTPrinter: Print SynthesizedProtocolAttrs when printing inheritance clauses
Imported types no longer have redundant inheritance clause
entries, because that messed up the ConformanceLookupTable,
so we have to print them explicitly here.

It would be better to always print the superclass, local
protocols and raw type and never look at the inheritance
clause at all, but that perturbs too many tests, and also
means we lose sugar (such as Swift 3/4 renaming aliases),
so I'm not going to do that now.