Remove benchmark CMakeLists and main.swift from gyb.
When a benchmark is added, it's name will be manually added to CMakeLists and
main.swift.

Adding a benchmark no longer requires discovering a nonobvious script.

If the build works locally, then it will work everywhere else.

Note that generate_harness.py should now be renamed to generate_benchmarks.py,
but I think it's name may be hard-coded somewhere in the compiler tests. Anyway,
this script should go away as soon as someone fixes the cmake configuration to
automatically run gyb.