[SILGen] Generate a trap for unexpected cases in all @objc enums
(both C enums and Swift enums declared @objc), even if declared
"exhaustive", because of the "feature" in C of treating a value not
declared as a case as a valid value of an enum.  No more undefined
behavior here!

This also adds more tests for trapping on unexpected cases for
non-exhaustive non-@objc enums.