Make update-checkout.py non-executable
utils/update_checkout.py was marked as executable.  This made it show
up in the bash auto-completion.  Worse, if one were to execute it, the
shell would try and interpret it as a shell script (since there is no
shebang line telling the shell that it is a Python script) and fail.

This commit removes the executable bits from file mode. This should
have been done when the original utils/update-checkout Python script
was renamed to update_checkout.py in dfe3af81.  That commit removed
the shebang but did not turn off the corresponding bits.