[stdlib] [prototype] Add new String comparison prototype.
Add in a new prototype, with integrated tests and benchmarks,
demonstrating an alternative approach to comparing Strings. Under the
prototype, the new ordering of Strings is the lexicographical ordering
of NFC-normalized UTF16 code units.

The prototype includes many fast paths and the high level algorithm
should be appropriate for most user strings. Speed ups depend on the
shape of the strings being compared, but it's not uncommon to see
2-10x improvement over Swift 4. There's plenty of low hanging fruit
left in the prototype to speed this up more, especially in common
scenarios.

The prototype currently does not have support for non-contiguous
NSStrings, which while normally ASCII in practice (99%+), can in
theory be another pathological form to support.