[CMake] Cleanup Swift library tablegen dependencies
In Clang with r309390 the tablegen headers are all wrapped by a target named clang-tablegen-targets, which is created as an empty target in Clang's CMake configuration. This allows depending on this target directly in both in and out-of-tree builds.

Similarly in LLVM r309389, intrinsics_gen is created in the CMake configuration so that out-of-tree projects can add it as a dependency.