[benchmark] Simplify how we filter tests down to 1 filter implementation.
Previously, we both filtered tests and marked tests as not being able to be run.
This commit rips out the marking of tests as being unable to be run (but counted
as part of the indices output by the compiler).

This also ensures benchmarks that are run are always given an ascending index
instead of sometimes having indices being skipped.