[ClangImporter] Prefer later enum_extensibility over earlier ones
Since NS_ENUM has an enum_extensibility(open) in it already, we want
to allow people to undo that by sticking enum_extensibility(closed) on
the end of their enum. Furthermore, we actually want to ignore the
enum_extensibility(open) that's in there now, so as not to introduce
new warnings in Swift 4.1.