Requirement environment: adopt the conformance's generic signature when we can.
Noticed by Slava: when the requirement is non-generic, the requirement
signature's generic signature is always going to be the same as the
conformance declaration context's generic signature, so use that immediately.

This eliminates a large number of generic signature builders created due to
requirement signatures (e.g., from 4893 -> 569 when measured against
the standard library work for SE-0157).

Short-circuit build of requirement environment.

This is all Slava and it’s magnificent.

FIXUP for “the Slava change”