[CMake] Default to Debug builds
When CMake is invoked manually to configure a Swift build, it fails
because no default value is specified for the build type:

    CMake Error at cmake/modules/SwiftUtils.cmake:94 (message):
      Unknown build type:
    Call Stack (most recent call first):
      CMakeLists.txt:98 (is_build_type_optimized)

Instead we take a cue from LLVM, and specify a default value for
CMAKE_BUILD_TYPE.  By providing a default, we reduce the number of
mandatory CMake configuration parameters, and make it easier for
people to get started.

The conditional was taken from llvm/CMakeLists.txt.  The check for
CMAKE_CONFIGURATION_TYPES is necessary to account for
multi-configuration generators [1].

The default was taken to be "Debug", simply because it was the one
which LLVM used.

[1]: https://cmake.org/pipermail/cmake/2006-February/008371.html